task 4

+++ Ch 4/input_numbers_from_console/input_numbers_from_console/main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 void main()
 {
-	int input1, input2;
+	double input1, input2;
 	while (cin >> input1 >> input2)
 	{
 		if (input1 > input2)
